requirements.txt
Django>=4.2.0
celery>=5.3.0
django-celery-results>=2.5.0
django-celery-beat>=2.5.0
redis>=4.6.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
